Added activity switching
Added methods for buttons on the welcome, login, and main screens to
move to other activities appropriately. The login does not check for a
valid login. The logout does not reset the current user.

Also note that the onCreate() method for LoginActivity is modified to
not set the login button to attemptLogin() as default for a basic login
activity.