Changed back button behavior
Back button on login screen calls finish() instead of creating a new
instance of WelcomeActivity so as not to create an overly long back
stack. Logging out clears the back stack up to the welcome activity so
that a user cannot go back to the main activity by pressing back on the
login screen.